+++ webui.bat
@@ -1,0 +1,1 @@
+python webui.py
