+++ webui.py
@@ -2,24 +2,14 @@ import os
 import argparse
 import numpy as np
 import gradio as gr
-from PIL import Image
 from interpolate_engine import InterpolateEngine
 from interpolate import Interpolate
 from collections import namedtuple
 from simple_log import SimpleLog
 from simple_config import SimpleConfig
 from auto_increment_filename import AutoIncrementFilename
-
-# idea make GIF to show the output
-
-def create_gif(images : dict, filepath : str, duration : int = 1000):
-    if len(images) < 1:
-        return None 
-    images = [Image.open(image) for image in images]
-    if len(images) == 1:
-        images[0].save(filepath)
-    else:
-        images[0].save(filepath, save_all=True, append_images=images[1:], optimize=False, duration=duration, loop=0)
+from image_utils import create_gif
+from file_utils import create_directories
 
 def interpolate(img_before_file : str, img_after_file : str):
     global log, config, engine
@@ -50,7 +40,7 @@ def main():
 
     log = SimpleLog(args.verbose)
     config = SimpleConfig(args.config_path).config_obj()
-    init_directories(config.output_dirs)
+    create_directories(config.output_dirs)
     engine = InterpolateEngine(config.model, config.gpu_ids)
 
     app = create_ui()
@@ -59,19 +49,23 @@ def main():
                 server_port = config.server_port)
 
 def create_ui():
-    with gr.Blocks() as app:
+    global config
+    with gr.Blocks(analytics_enabled=False, 
+                    title="VFIformer Web UI", 
+                    theme=config.user_interface["theme"],
+                    css=config.user_interface["css_file"]) as app:
         gr.Markdown("VFIformer Web UI")
         with gr.Tab("Frame Interpolation"):
-            with gr.Row():
-                with gr.Column():
+            with gr.Row(variant="compact"):
+                with gr.Column(variant="panel"):
                     img1_input = gr.Image(type="filepath", label="Before Image") #.style(height=300, width=400)
                     img2_input = gr.Image(type="filepath", label="After Image") #.style(height=300, width=400)
-                with gr.Column():
+                    interpolate_button = gr.Button("Interpolate", variant="primary")
+                with gr.Column(variant="panel"):
                     img_output = gr.Image(type="filepath", label="Animated Preview") #.style(height=300, width=400)
                     file_output = gr.File(type="file", label="Download Interpolated Image")
-            interpolate_button = gr.Button("Interpolate", variant="primary")
         with gr.Tab("Slow Motion"):
-            with gr.Row():
+            with gr.Row(variant="compact"):
                 image_input = gr.Image()
                 image_output = gr.Image()
             image_button = gr.Button("Flip")
@@ -79,13 +73,6 @@ def create_ui():
         interpolate_button.click(interpolate, inputs=[img1_input, img2_input], outputs=[img_output, file_output])
         # image_button.click(flip_image, inputs=image_input, outputs=image_output)
     return app
-
-def init_directories(dirs : dict):
-    for key in dirs.keys():
-        dir = dirs[key]
-        if not os.path.exists(dir):
-            log.log(f"creating output directory {dir}")
-            os.makedirs(dir)
 
 if __name__ == '__main__':
     main()
